% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpSig.R
\name{computeSim}
\alias{computeSim}
\title{Compute similarity between two mutation data.}
\usage{
computeSim(mut1, mut2)
}
\arguments{
\item{mut1}{Input sample mutation data.}

\item{mut2}{Refrence mutation data.}
}
\value{
Score of similarity
}
\description{
Compute similarity between two mutation data.
}
\keyword{internal}
