% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mSigdb}
\alias{mSigdb}
\title{Values for 28 Cancer project}
\format{A data frame of 32 rows and 73 columns}
\description{
A dataset containing the value for 28 projects that sort out by mSignaturesdb
}
\keyword{datasets}
