% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mSigsitedb}
\alias{mSigsitedb}
\title{Values for 65 Cancer project}
\format{A data frame of 32 rows and 65 columns}
\description{
A dataset containing the value for 65 projects that sort out by mSignaturesdb
}
\keyword{datasets}
