% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sample.clinical}
\alias{sample.clinical}
\title{Values for 2 samples}
\format{A data frame of 2 rows and 3 columns}
\description{
A dataset containing the clinical data for 2 samples
}
\keyword{datasets}
