% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sample.mut.ref}
\alias{sample.mut.ref}
\title{Example input to transMut()}
\format{A data frame of 657 rows and 5 columns}
\description{
A data frame containing example mutations that can be used as input to transMut()
Data comes from two TCGA LUAD patients
}
\keyword{datasets}
