% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sample.mut.trans}
\alias{sample.mut.trans}
\title{Values for 2 randomly generated tumor samples
Example input to sigSim()}
\format{A data frame of 2 rows and 96 columns}
\description{
A dataset containing the value for the 96-trinucleotide contexts of 100 tumors
}
\keyword{datasets}
