% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpSig.R
\name{transMut}
\alias{transMut}
\title{Transform the input data to simSig input data}
\usage{
transMut(tmMut, bsg = NULL)
}
\arguments{
\item{tmMut}{Location of the mutation file that is to be converted or name of data frame in environment.
Use data(sample.mut.ref) to see the example data.}

\item{bsg}{Location of the mutation file that is to be converted or name of data frame in environment.
The default value is BSgenome.Hsapiens.UCSC.hg19, and Only set if another genome build is required. Must be a BSgenome object.}
}
\value{
A data frame that contains sample IDs for the rows and trinucleotide contexts for the columns.
Each entry is the count of how many times a mutation with that trinucleotide context is seen in the sample.
}
\description{
Transform the input data to simSig input data
}
\examples{
testTrans = transMut(tmMut = sample.mut.ref)
}
